requirements.txt
asgiref==3.7.1
async-timeout==4.0.2
Django==4.2.1
django-cache-url==3.4.4
django-cors-headers==4.0.0
django-countries==7.5.1
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
PyJWT==1.7.1
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
sqlparse==0.4.4
typing_extensions==4.6.1
